requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
gunicorn==21.2.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==42.0.0

# AI/ML Dependencies
openai==1.10.0
transformers==4.37.2
torch==2.2.0
torchvision==0.17.0

# Image Processing
Pillow==10.2.0
opencv-python-headless==4.9.0.80
numpy==1.26.3

# HTTP & Caching
httpx==0.26.0
requests==2.31.0
redis==5.0.1

# Monitoring & Logging
sentry-sdk==1.40.0
prometheus-client==0.19.0

# Utilities
python-dateutil==2.8.2
pytz==2024.1

# Development (optional)
pytest==8.0.0
pytest-asyncio==0.23.3
black==24.1.1
flake8==7.0.0
mypy==1.8.0

# Streamlit (for frontend)
streamlit==1.30.0

# AWS S3 (optional, for cloud storage)
boto3==1.34.34

# Email (optional)
aiosmtplib==3.0.1
